requirements.txt
Flask
dataclasses-json
PyMySQL
SQLAlchemy
yoyo-migrations
flask_swagger_ui
